component-ise the "play audio button"
