+++ www/app/components/index.ts
@@ -1,0 +1,6 @@
+'use strict';
+
+var angular = require('angular');
+
+angular.module('app')
+    .component('playerButton', require('./player-button.component'))
