+++ www/app/components/player-button.component.html
@@ -1,0 +1,7 @@
+<button class="ms-Button ms-Button--hero " ng-click="$ctrl.play()">
+    <span class="ms-Button-icon">
+        <i ng-hide="$ctrl.isAudioPlaying" class="ms-Icon ms-Icon--play"></i>
+        <i ng-show="$ctrl.isAudioPlaying" class="ms-Icon ms-Icon--ellipsis"></i>
+    </span> 
+    <span class="ms-Button-label">{{$ctrl.text}}</span>
+</button>
