+++ www/webpack.config.js
@@ -25,6 +25,7 @@ module.exports = {
     module: {
         rules: [
             { test: /\.(t|j)s$/, use: { loader: 'awesome-typescript-loader?{tsconfig: "tsconfig.json"}' } },
+            { test: /\.(html)$/, use: { loader: 'html-loader'} },
             // addition - add source-map support 
             { enforce: "pre", test: /\.js$/, loader: "source-map-loader" }
         ]
